% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Function to check convergence of estimation model}
\usage{
check_convergence(em_names, om_sim_num, col, plot_ncol, plot_nrow, casedir)
}
\arguments{
\item{em_names}{Names of estimation models}

\item{om_sim_num}{Number of iterations from the operating model}

\item{col}{Color vector to specify colors for operating model and each estimation model}

\item{plot_ncol}{Number of columns in a comparison figure}

\item{plot_nrow}{Number of rows in a comparison figure}

\item{casedir}{Case working directory}
}
\description{
Function to check convergence of estimation model
}
