% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_f_case.R
\name{f_case}
\alias{f_case}
\title{Compute fishing mortality rate over years.}
\usage{
f_case(
  f_pattern = 1,
  start_val = 0.01,
  middle_val = 0.39,
  end_val = NULL,
  f_val = NULL,
  start_year = NULL,
  middle_year = NULL,
  nyr = 30,
  om_sim_num = 160,
  f_dev_matrix = f_dev_matrix,
  initial_equilibrium_F = TRUE
)
}
\arguments{
\item{f_pattern}{A pattern of fishing mortality rate.
f_pattern: 1. increase
2. decrease
3. increase first, then decrease
4. decrease first, then increase
5. constantly fluctuate around a value
6. customized pattern}

\item{start_val}{start value of the main shape of fishing mortality}

\item{middle_val}{middle value of the main shape of fishing mortality for case 3 and 4}

\item{end_val}{end value of the main shape of fishing mortality for case 3 and 4}

\item{f_val}{customized annual value of the main shape of fishing mortality}

\item{start_year}{start year for the start value of fishing mortality}

\item{middle_year}{middle year for the middle value of fishing mortality}

\item{nyr}{totla number of years}

\item{om_sim_num}{total iteration numbers}

\item{f_dev_matrix}{F deviations-at-age per iteration}
}
\value{
A matrix of fishing mortality rate \code{f_matrix} over years (column) for each iteartion (row)
}
\description{
Compute fishing mortality rate over years.
}
\examples{
f_case(f_pattern=1, start_val=0.01, middle_val=NULL, end_val=0.39, f_val=NULL, start_year=1, middle_year=NULL, nyr=30, om_sim_num=160, f_dev_matrix=f_dev_matrix)
f_case(f_pattern=3, start_val=0.01, middle_val=0.39, end_val=0.19, f_val=NULL, start_year=1, middle_year=24, nyr=30, om_sim_num=160, f_dev_matrix=f_dev_matrix)
f_case(f_pattern=5, start_val=0.01, middle_val=NULL, end_val=NULL, f_val=NULL, start_year=1, middle_year=NULL, nyr=30, om_sim_num=160, f_dev_matrix=f_dev_matrix)
f_case(f_pattern=6, f_val=rnorm(n=30, mean=1, sd=0.3), nyr=30, om_sim_num=160, f_dev_matrix=f_dev_matrix)
}
