% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_deviation_generator.R
\name{f_dev_case}
\alias{f_dev_case}
\title{F deviations generator}
\usage{
f_dev_case(
  f_dev_change,
  nyr = 30,
  logf_sd = 0.1,
  om_sim_num = 160,
  f_dev_sum2zero = FALSE
)
}
\arguments{
\item{f_dev_change}{if True, generate different F deviations per iteration; If FALSE, generate same F deviations per iteration}

\item{nyr}{number of years}

\item{om_sim_num}{number of iterations per case}

\item{f_dev_sum2zero}{Sum F deviations to 0 or not}

\item{logF.sd}{SD of fishing mortality in log space}
}
\description{
F deviations generator
}
