% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R
\name{generate_plot}
\alias{generate_plot}
\title{Function to generate comparison figures}
\usage{
generate_plot(
  em_names = NULL,
  plot_ncol = NULL,
  plot_nrow = NULL,
  plot_color = NULL,
  input_list = NULL,
  adhoc_bias_cor = FALSE
)
}
\arguments{
\item{plot_ncol}{Number of columns in a comparison figure}

\item{plot_nrow}{Number of rows in a comparison figure}

\item{plot_color}{Color vector to specify colors for operating model and each estimation model}

\item{input_list}{Name of the input data list}

\item{adhoc_bias_cor}{Use ad hoc bias correction in recruitment? The default setting is FALSE.}
}
\description{
Function to compare OM true values and EM estimates and generate comparison figures
}
