% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{geomean}
\alias{geomean}
\title{Compute geometric mean of a time series x.}
\usage{
geomean(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The geomean: \code{y}.
}
\description{
Compute geometric mean of a time series x.
}
\examples{
geomean(x=c(1,2,3,4,5))
geomean(x=rnorm(n=100, mean=100, sd=30))
}
