% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_deviation_generator.R
\name{r_dev_case}
\alias{r_dev_case}
\title{Recruitment deviations generator}
\usage{
r_dev_case(
  r_dev_change = TRUE,
  nyr = 30,
  logR_sd = 0.1,
  om_sim_num = 160,
  r_dev_sum2zero = FALSE
)
}
\arguments{
\item{nyr}{number of years}

\item{om_sim_num}{number of iterations per case}

\item{r_dev_sum2zero}{Sum recruitment deviations to 0 or not}

\item{rec_dev_change}{if True, generate different recruitment deviations per iteration; If FALSE, generate same recruitment deviations per iteration}

\item{logR.sd}{recruitment SD in log space}
}
\description{
Recruitment deviations generator
}
