% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plot_data.R
\name{read_plot_data}
\alias{read_plot_data}
\title{Function to read key output from OM and EMs}
\usage{
read_plot_data(
  em_names = NULL,
  casedir = NULL,
  keep_sim_num = NULL,
  adhoc_bias_cor = FALSE,
  SRmodel = 1
)
}
\arguments{
\item{em_names}{Names of estimation models}

\item{casedir}{Case working directory}

\item{keep_sim_num}{Final number of iterations saved for a case}

\item{adhoc_bias_cor}{Use ad hoc bias correction in recruitment? The default setting is FALSE.}

\item{SRmodel}{Spawner-recruit model. 1 = Beverton-Holt model; 2 = Ricker model}
}
\description{
Function to read key output from OM and EMs
}
