% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fims.R
\name{run_fims}
\alias{run_fims}
\title{Function to run Fisheries Integrated Modeling System (https://github.com/NOAA-FIMS/FIMS)}
\usage{
run_fims(
  maindir = maindir,
  subdir = "FIMS",
  om_sim_num = NULL,
  casedir = casedir,
  em_bias_cor = em_bias_cor
)
}
\arguments{
\item{maindir}{Main working directory}

\item{subdir}{Estimation model working directory}

\item{om_sim_num}{Number of iterations from the operating model}

\item{casedir}{Case working directory}

\item{em_bias_cor}{Use bias correction in the estimation model or not?}
}
\description{
Function to read simulated true values from the operating model and to run FIMS
}
