% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mas.R
\name{run_mas}
\alias{run_mas}
\title{Function to run Metapopulation Assessment System (https://github.com/nmfs-fish-tools/r4MAS)}
\usage{
run_mas(
  maindir = maindir,
  subdir = "MAS",
  om_sim_num = NULL,
  casedir = casedir
)
}
\arguments{
\item{maindir}{Main working directory}

\item{subdir}{Estimation model working directory}

\item{om_sim_num}{Number of iterations from the operating model}

\item{casedir}{Case working directory}
}
\description{
Function to read simulated true values from the operating model and to run MAS
}
