% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_om.R
\name{run_om}
\alias{run_om}
\title{Run operating model with values from \code{input_list}}
\usage{
run_om(input_list = NULL, show_iter_num = FALSE)
}
\description{
Run operating model with values from \code{input_list}
}
