% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_initial_input.R
\name{save_initial_input}
\alias{save_initial_input}
\title{Save initial input values to a list}
\usage{
save_initial_input(base_case = TRUE, input_list = NULL, maindir = getwd(), ...)
}
\arguments{
\item{base_case}{A logical value to declare if you want the default values
that are stored inside this function for case 1, i.e., \code{"C1"}.}

\item{input_list}{If \code{base_case = FALSE}, then \code{input_list} must be provided.
The argument can take any list but a list previously returned from
\code{save_initial_input()} is ideal such that it has all of the necessary
elements.}

\item{maindir}{The directory where you the results from down-stream
functions like \code{\link[=run_OM]{run_OM()}} will save the results to.
TODO - decide if this argument should be added and how to force it to
make changes because it will not be included in ...
- decide if we should use dir.create() to ensure that the directory
exists
- is the current directory okay as a default or should it be
a temporary directory?}

\item{...}{Elements of \code{input_list} or the returned list using the base-case
scenario if \code{base_case = TRUE} that you want to change.}
}
\value{
A list is object is returned. Each element is necessary to run a
simulation via passing the list to \code{\link[=run_OM]{run_OM()}}.
TODO - use the saved data object to generate a bulleted list of names(C1).
- double check all of the input values
}
\description{
The default is to return a list containing the values needed to
run case one, i.e., \code{"C1"}. This function can also be used to modify
the list elements using \code{...} rather than assigning each change with \verb{<-}.
}
\examples{
# Create the default input for case one, i.e., "C1"
null_case_input <- save_initial_input()
# Change the default to case two, i.e., "C2"
updated_input <- save_intial_input(
  base_case = FALSE,
  input_list = null_case_input,
  case_name = "C2",
  logR_sd = 0.6
)

}
\seealso{
\code{\link[=run_OM]{run_OM()}}
}
\author{
Bai Li
}
